Fixed bugs

+++ app/Http/Controllers/CategoryController.php
@@ -11,23 +11,42 @@ use App\Helpers\DatabaseConnection;
 
 class CategoryController extends Controller
 {
-    public function list()
+    public function list(Request $request)
     {
         try {
             DatabaseConnection::setConnection();
-            $categories = Category::all();
+            Category::where('type_id', null)->forceDelete();
+            if ($request->id > 0) {
+                $categories = Category::where('up_category_id', $request->id)->get();
+                $up_category = Category::find($request->id);
+                if ($up_category->up_category_id == null)
+                    $up_category = null;
+                else
+                    $up_category = Category::find($up_category->up_category_id);
+            } else {
+                $categories = Category::where('main_category', BooleanEnum::fromValue(1))->get();
+                $up_category = null;
+            }
         } catch (\Exception $e) {
             $categories = null;
+            $up_category = null;
         }
         return view('client.category.list', compact([
             'categories', $categories,
+            'up_category', $up_category,
         ]));
     }
 
-    public function create()
+    public function create(Request $request)
     {
         DatabaseConnection::setConnection();
+        if($request->up_category_id) {
+            $up_category_id = Category::find($request->up_category_id)->up_category_id;
+        }else{
+            $up_category_id = null;
+        }
         $category = Category::create([
+            'up_category_id' => $up_category_id,
             'title' => 'Yeni Kategori',
         ]);
         return $category->id;
@@ -37,15 +56,39 @@ class CategoryController extends Controller
     {
         DatabaseConnection::setConnection();
         $category = Category::find($request->id);
-        if ($request->can_sub_category_id == 'false')
-            $can_sub_category_id = 0;
-        elseif ($request->can_sub_category_id == 'true')
-            $can_sub_category_id = 1;
+
+
+        if ($request->can_sub_category == 'false') {
+            $sub_categories = Category::where('up_category_id', $category->id)->get();
+            $messages = [
+                'status' => 'warning',
+                'title' => 'Kaydedilemedi',
+                'message' => 'Mevcut alt kategoriler bulundu.',
+            ];
+            if (count($sub_categories) > 0) {
+                return response()->json(['messages' => $messages, 'sub_categories' => $sub_categories]);
+            }
+            $can_sub_category = 0;
+        } elseif
+        ($request->can_sub_category == 'true')
+            $can_sub_category = 1;
+
+
+        if ($request->main_category == 'false') {
+            $main_category = 0;
+            $up_category_id = $request->up_category_id;
+        } elseif ($request->main_category == 'true') {
+            $main_category = 1;
+            $up_category_id = null;
+        }
+
+
         if ($category->update([
-            'up_category_id' => $request->up_category_id,
+            'up_category_id' => $up_category_id,
             'type_id' => CategoryType::fromValue(CategoryType::parseDatabase($request->type_id)),
             'title' => $request->title,
-            'can_sub_category_id' => BooleanEnum::fromValue(BooleanEnum::parseDatabase($can_sub_category_id)),
+            'can_sub_category' => BooleanEnum::fromValue(BooleanEnum::parseDatabase($can_sub_category)),
+            'main_category' => BooleanEnum::fromValue(BooleanEnum::parseDatabase($main_category)),
         ])) {
             $messages = [
                 'status' => 'success',
@@ -67,7 +110,7 @@ class CategoryController extends Controller
         DatabaseConnection::setConnection();
         $selected_category = Category::find($request->id);
         $selected_up_category = Category::find($selected_category->up_category_id);
-        $categories = Category::where('can_sub_category_id', BooleanEnum::fromValue(1))->where('type_id', $request->type_id)->get()->except([$selected_category->id]);
+        $categories = Category::where('can_sub_category', BooleanEnum::fromValue(1))->where('type_id', $request->type_id)->get()->except([$selected_category->id]);
 
         return response()->json([
             'categories' => $categories,
@@ -94,34 +137,43 @@ class CategoryController extends Controller
     {
         DatabaseConnection::setConnection();
         $category = Category::find($request->id);
-        if ($category) {
-            if ($category->title === "Yeni Kategori") {
-                if ($category->forceDelete()) {
+        $sub_categories = Category::where('up_category_id', $category->id)->get();
+        if (count($sub_categories) > 0) {
+            $messages = [
+                'status' => 'warning',
+                'title' => 'Silinemedi',
+                'message' => 'Alt kategoriler bulundu. Önce alt kategorileri silmeniz gerekmektedir.',
+            ];
+        } else {
+            if ($category) {
+                if ($category->title === "Yeni Kategori") {
+                    if ($category->forceDelete()) {
+                        $messages = [
+                            'status' => 'success',
+                            'title' => 'Silindi',
+                            'message' => 'Yönlendiriliyorsunuz.',
+                        ];
+                    }
+                } else if ($category->delete()) {
                     $messages = [
                         'status' => 'success',
                         'title' => 'Silindi',
                         'message' => 'Yönlendiriliyorsunuz.',
                     ];
+                } else {
+                    $messages = [
+                        'status' => 'warning',
+                        'title' => 'Silinemedi',
+                        'message' => 'Yönlendiriliyorsunuz.',
+                    ];
                 }
-            } else if ($category->delete()) {
-                $messages = [
-                    'status' => 'success',
-                    'title' => 'Silindi',
-                    'message' => 'Yönlendiriliyorsunuz.',
-                ];
             } else {
                 $messages = [
                     'status' => 'warning',
                     'title' => 'Silinemedi',
-                    'message' => 'Yönlendiriliyorsunuz.',
+                    'message' => 'Kategori bulunamadı.',
                 ];
             }
-        } else {
-            $messages = [
-                'status' => 'warning',
-                'title' => 'Silinemedi',
-                'message' => 'Kategori bulunamadı.',
-            ];
         }
         return response()->json([
             'messages' => $messages,
